+++ inst/tinytest/ctrdata_euctr.R
@@ -491,6 +491,7 @@ expect_false(
 expect_true(
   all(c("NCT00001209", "NCT00001436", "NCT00187109", "NCT01516567") %in%
         trialsEuctr))
+
 # test
 expect_false(
   any(c("NCT01471782") %in% trialsEuctr))